+++ config/constants.ts
@@ -112,30 +112,17 @@ export const icav2AccessTokenSecretName: Record<AppStage, string> = {
 };
 
 /*
-Resources generated by the ICAv2 Copy Batch utility stack
- */
-export const icav2CopyBatchSSMRoot = '/orcabus/icav2_copy_batch_utility';
-export const icav2CopyBatchSSMBatchArn = path.join(icav2CopyBatchSSMRoot, 'batch_sfn_arn');
-export const icav2CopyBatchSSMBatchName = path.join(icav2CopyBatchSSMRoot, 'batch_sfn_name');
-export const icav2CopyBatchSSMSingleArn = path.join(icav2CopyBatchSSMRoot, 'single_sfn_arn');
-export const icav2CopyBatchSSMSingleName = path.join(icav2CopyBatchSSMRoot, 'single_sfn_name');
-export const icav2CopyBatchUtilityName = 'icav2_copy_batch_utility_sfn';
-export const icav2CopySingleUtilityName = 'icav2_copy_single_utility_sfn';
-
-/*
 Resources generated by the BSSH Fastq Copy Manager
- */
+*/
 
-export const bsshFastqCopyManagerSSMRoot = '/orcabus/bssh_icav2_fastq_copy_manager';
-export const bsshFastqCopyManagerSSMName = path.join(
-  bsshFastqCopyManagerSSMRoot,
-  'state_machine_name'
-);
-export const bsshFastqCopyManagerSSMArn = path.join(
-  bsshFastqCopyManagerSSMRoot,
-  'state_machine_arn'
-);
-export const bsshFastqCopyManagerSfnName = 'bssh_fastq_copy_manager_sfn';
+export const bsshFastqCopyManagerWorkflowType = 'bssh_fastq_copy';
+export const bsshFastqCopyManagerWorkflowTypeVersion = '1.0.0';
+export const bsshFastqCopyManagerServiceVersion = '2024.05.15';
+
+export const bsshFastqCopyManagerReadyEventSource = 'orcabus.wfm';
+export const bsshFastqCopyManagerEventSource = 'orcabus.bssh_fastq_copy_manager';
+export const bsshFastqCopyManagerEventDetailType = 'workflowRunStateChange';
+export const bsshFastqCopyManagerSfnPrefix = 'bsshFastqCopyManagerSfn';
 
 /*
 Resources generated by the BCLConvert InterOp QC pipeline
@@ -157,14 +144,6 @@ export const bclconvertInteropQcDynamoDbTableSSMArn = path.join(
 );
 
 // Stateless
-export const bclconvertInteropQcIcav2PipelineSfnSSMName = path.join(
-  bclConvertInteropQcSSMRoot,
-  'state_machine_name'
-);
-export const bclconvertInteropQcIcav2PipelineSfnSSMArn = path.join(
-  bclConvertInteropQcSSMRoot,
-  'state_machine_arn'
-);
 
 export const bclconvertInteropQcIcav2PipelineWorkflowType = 'bclconvert_interop_qc';
 export const bclconvertInteropQcIcav2PipelineWorkflowTypeVersion = '1.3.1--1.21';
@@ -201,16 +180,6 @@ export const cttsov2DynamoDbTableSSMName = path.join(
 export const cttsov2DynamoDbTableSSMArn = path.join(
   cttsov2Icav2PipelineManagerSSMRoot,
   'dynamodb_table_arn'
-);
-
-// Stateless
-export const cttsov2Icav2PipelineSfnSSMName = path.join(
-  cttsov2Icav2PipelineManagerSSMRoot,
-  'state_machine_name'
-);
-export const cttsov2Icav2PipelineSfnSSMArn = path.join(
-  cttsov2Icav2PipelineManagerSSMRoot,
-  'state_machine_arn'
 );
 
 // Event Handling
