+++ lib/workload/stateless/stacks/bssh-icav2-fastq-copy-manager/deploy/constructs/bssh-icav2-fastq-copy-manager/index.ts
@@ -12,22 +12,30 @@ import { PythonFunction } from '@aws-cdk/aws-lambda-python-alpha';
 import * as secretsmanager from 'aws-cdk-lib/aws-secretsmanager';
 
 interface BsshIcav2FastqCopyStateMachineConstructProps {
+  /* Stack objects */
   icav2CopyBatchStateMachineObj: sfn.IStateMachine;
-  icav2JwtSsmParameterObj: secretsmanager.ISecret;
+  icav2JwtSecretObj: secretsmanager.ISecret;
   lambdasLayerObj: lambda.ILayerVersion;
   eventBusObj: events.IEventBus; // Orcabus main event
+  /* Lambda paths */
   bclconvertSuccessEventHandlerLambdaPath: string; // __dirname + '/../../../lambdas/bclconvert_success_event_handler'
-  // execution_check_status_lambda_path: string; // __dirname + '../lambdas/check_execution_completion'
+  /* State machine */
+  stateMachineName: string; // 'bssh-fastq-copy-sfn '
   workflowDefinitionBodyPath: string; // __dirname + '../step_functions_templates/bclconvert_success_event_state_machine.json'
+  // Event handling //
+  detailType: string; // workflowRunStateChange
+  serviceVersion: string; // 2024.05.15
+  triggerLaunchSource: string; // orcabus.wfm
+  internalEventSource: string; // orcabus.bssh_fastq_copy
+  workflowType: string; // bssh_fastq_copy
+  workflowVersion: string; // 1.0.0
 }
 
 export class BsshIcav2FastqCopyStateMachineConstruct extends Construct {
-  public readonly icav2BclconvertSuccessEventSsmStateMachineObj: sfn.IStateMachine;
-
   constructor(scope: Construct, id: string, props: BsshIcav2FastqCopyStateMachineConstructProps) {
     super(scope, id);
 
-    // Workflow Session lambda
+    // Handle bclconvert success lambda
     const bclconvert_success_event_lambda = new PythonFunction(
       this,
       'bclconvert_success_event_lambda_python_function',
@@ -42,13 +50,15 @@ export class BsshIcav2FastqCopyStateMachineConstruct extends Construct {
         timeout: Duration.seconds(60),
         environment: {
           ICAV2_BASE_URL: 'https://ica.illumina.com/ica/rest',
-          ICAV2_ACCESS_TOKEN_SECRET_ID: props.icav2JwtSsmParameterObj.secretName,
+          ICAV2_ACCESS_TOKEN_SECRET_ID: props.icav2JwtSecretObj.secretName,
         },
       }
     );
 
     // Add icav2 secrets permissions to lambda
-    props.icav2JwtSsmParameterObj.grantRead(<iam.IRole>bclconvert_success_event_lambda.role);
+    props.icav2JwtSecretObj.grantRead(
+      <iam.IRole>bclconvert_success_event_lambda.currentVersion.role
+    );
 
     // Specify the statemachine and replace the arn placeholders with the lambda arns defined above
     const stateMachine = new sfn.StateMachine(this, 'bssh_fastq_copy_state_machine', {
@@ -56,19 +66,17 @@ export class BsshIcav2FastqCopyStateMachineConstruct extends Construct {
       definitionBody: DefinitionBody.fromFile(props.workflowDefinitionBodyPath),
       // definitionSubstitutions
       definitionSubstitutions: {
-        __bclconvert_success_event_lambda_arn__: bclconvert_success_event_lambda.functionArn,
+        __bclconvert_success_event_lambda_arn__:
+          bclconvert_success_event_lambda.currentVersion.functionArn,
         __copy_batch_data_state_machine_arn__: props.icav2CopyBatchStateMachineObj.stateMachineArn,
         __eventbus_name__: props.eventBusObj.eventBusName,
       },
     });
 
-    // Add execution permissions to stateMachine role
-    stateMachine.addToRolePolicy(
-      new iam.PolicyStatement({
-        resources: [bclconvert_success_event_lambda.functionArn],
-        actions: ['lambda:InvokeFunction'],
-      })
-    );
+    // Add lambda invoke permissions to stateMachine role
+    bclconvert_success_event_lambda.currentVersion.grantInvoke(stateMachine.role);
+
+    // Allow the icav2 copy batch statemachine to be started by the bssh fastq copy manager
 
     // Because we run a nested state machine, we need to add the permissions to the state machine role
     // See https://stackoverflow.com/questions/60612853/nested-step-function-in-a-step-function-unknown-error-not-authorized-to-cr
@@ -81,15 +89,17 @@ export class BsshIcav2FastqCopyStateMachineConstruct extends Construct {
       })
     );
 
+    // State machine
+    props.icav2CopyBatchStateMachineObj.grantStartExecution(stateMachine.role);
+
     // Trigger state machine on event
     const rule = new events.Rule(this, 'bssh_fastq_copy_trigger_rule', {
       eventBus: props.eventBusObj,
       eventPattern: {
-        // FIXME - nothing set in stone yet
-        source: ['orcabus.bssh'],
-        detailType: ['BsshBclconvertStateChange'],
+        source: [props.triggerLaunchSource],
+        detailType: [props.detailType],
         detail: {
-          status: ['succeeded'],
+          status: ['ready'],
         },
       },
     });
@@ -103,12 +113,5 @@ export class BsshIcav2FastqCopyStateMachineConstruct extends Construct {
 
     // Allow the statemachine to submit events to the event bus
     props.eventBusObj.grantPutEventsTo(stateMachine.role);
-
-    // Allow the icav2 copy batch statemachine to be started by the bssh fastq copy manager
-    // State machine
-    props.icav2CopyBatchStateMachineObj.grantStartExecution(stateMachine.role);
-
-    // Set outputs
-    this.icav2BclconvertSuccessEventSsmStateMachineObj = stateMachine;
   }
 }
