+++ lib/workload/stateless/stacks/bssh-icav2-fastq-copy-manager/deploy/stack.ts
@@ -7,37 +7,37 @@ import * as secretsmanager from 'aws-cdk-lib/aws-secretsmanager';
 import * as events from 'aws-cdk-lib/aws-events';
 import path from 'path';
 import { PythonLambdaLayerConstruct } from '../../../../components/python-lambda-layer';
+import { ICAv2CopyBatchUtilityConstruct } from '../../../../components/icav2-copy-files-batch';
 
 export interface BsshIcav2FastqCopyManagerConfig {
-  // Define construct properties here
-  icav2JwtSecretsManagerPath: string; // "/icav2/umccr-prod/service-production-jwt-token-secret-arn"
-  icav2CopyBatchUtilityStateMachineName: string; // "/icav2/umccr-prod/copy-batch-state-machine/name"
-  bsshIcav2FastqCopyManagerStateMachineName: string; // "bssh_icav2_fastq_copy_manager"
-  bsshIcav2FastqCopyManagerStateMachineNameSsmParameterPath: string; // "/bssh_icav2_fastq_copy/state_machine/name"
-  bsshIcav2FastqCopyManagerStateMachineArnSsmParameterPath: string; // "/bssh_icav2_fastq_copy/state_machine/arn"
+  /* Required external properties */
+  icav2TokenSecretId: string; // "ICAv2JWTKey-umccr-prod-service-trial"
+
+  /* Configurables for this stack */
+  // The name of this AWS Step function state machine
+  bsshIcav2FastqCopyManagerStateMachinePrefix: string; // "bssh_icav2_fastq_copy_manager"
+
+  /* Events Handling */
   eventBusName: string; // OrcabusMain
+  workflowType: string; // bssh_fastq_copy
+  workflowVersion: string; // 1.0.0
+  serviceVersion: string; // 2024.05.15
+  triggerLaunchSource: string; // orcabus.wfm
+  internalEventSource: string; // orcabus.bssh_fastq_copy
+  detailType: string; // workflowRunStateChange
 }
 
 export type BsshIcav2FastqCopyManagerStackProps = BsshIcav2FastqCopyManagerConfig & cdk.StackProps;
 
 export class BsshIcav2FastqCopyManagerStack extends cdk.Stack {
-  public readonly bsshIcav2FastqCopyStateMachineArn: string;
-
   constructor(scope: Construct, id: string, props: BsshIcav2FastqCopyManagerStackProps) {
     super(scope, id, props);
 
-    // Get the copy batch state machine value from lookup
-    const icav2_copy_batch_stack_state_machine_arn_obj = sfn.StateMachine.fromStateMachineName(
-      this,
-      'icav2_copy_batch_state_machine',
-      props.icav2CopyBatchUtilityStateMachineName
-    );
-
     // Get the icav2 jwt secret from lookup
-    const icav2_jwt_ssm_parameter_obj = secretsmanager.Secret.fromSecretNameV2(
+    const icav2_jwt_secret_obj = secretsmanager.Secret.fromSecretNameV2(
       this,
       'icav2_jwt_secret',
-      props.icav2JwtSecretsManagerPath
+      props.icav2TokenSecretId
     );
 
     // Get the lambda layer object
@@ -50,11 +50,22 @@ export class BsshIcav2FastqCopyManagerStack extends cdk.Stack {
     // Get eventbus object
     const event_bus_obj = events.EventBus.fromEventBusName(this, 'event_bus', props.eventBusName);
 
+    // Construct the icav2 copy files state machines
+    const icav2_copy_batch_state_machine = new ICAv2CopyBatchUtilityConstruct(
+      this,
+      'icav2_copy_batch_state_machine',
+      {
+        icav2JwtSecretParameterObj: icav2_jwt_secret_obj,
+        stateMachineNameSingle: `${props.bsshIcav2FastqCopyManagerStateMachinePrefix}-single-icav2-files-copy`,
+        stateMachineNameBatch: `${props.bsshIcav2FastqCopyManagerStateMachinePrefix}-batch-icav2-files-copy`,
+      }
+    );
+
     const icav2_bclconvert_success_event_state_machine =
       new BsshIcav2FastqCopyStateMachineConstruct(this, id, {
         /* Stack objects */
-        icav2CopyBatchStateMachineObj: icav2_copy_batch_stack_state_machine_arn_obj,
-        icav2JwtSsmParameterObj: icav2_jwt_ssm_parameter_obj,
+        icav2CopyBatchStateMachineObj: icav2_copy_batch_state_machine.icav2CopyFilesBatchSfnObj,
+        icav2JwtSecretObj: icav2_jwt_secret_obj,
         lambdasLayerObj: lambda_layer_obj.lambdaLayerVersionObj,
         eventBusObj: event_bus_obj,
         /* Lambda paths */
@@ -62,29 +73,19 @@ export class BsshIcav2FastqCopyManagerStack extends cdk.Stack {
           __dirname,
           '../lambdas/query_bclconvert_outputs_handler'
         ),
-        /* State machine paths */
+        /* State machine */
+        stateMachineName: props.bsshIcav2FastqCopyManagerStateMachinePrefix,
         workflowDefinitionBodyPath: path.join(
           __dirname,
           '../step_functions_templates/bclconvert_success_event_state_machine.json'
         ),
+        // Event handling //
+        detailType: props.detailType,
+        serviceVersion: props.serviceVersion,
+        triggerLaunchSource: props.triggerLaunchSource,
+        internalEventSource: props.internalEventSource,
+        workflowType: props.workflowType,
+        workflowVersion: props.workflowVersion,
       });
-
-    // Set outputs
-    this.bsshIcav2FastqCopyStateMachineArn =
-      icav2_bclconvert_success_event_state_machine.icav2BclconvertSuccessEventSsmStateMachineObj.stateMachineArn;
-
-    // Set ssm parameter paths
-    new ssm.StringParameter(
-      this,
-      'bssh_icav2_fastq_copy_manager_state_machine_name_ssm_parameter',
-      {
-        parameterName: props.bsshIcav2FastqCopyManagerStateMachineNameSsmParameterPath,
-        stringValue: props.bsshIcav2FastqCopyManagerStateMachineName,
-      }
-    );
-    new ssm.StringParameter(this, 'bssh_icav2_fastq_copy_manager_state_machine_arn_ssm_parameter', {
-      parameterName: props.bsshIcav2FastqCopyManagerStateMachineArnSsmParameterPath,
-      stringValue: this.bsshIcav2FastqCopyStateMachineArn,
-    });
   }
 }
