--- lib/workload/stateless/stacks/icav2-copy-batch-utility/.npmignore
@@ -1,6 +1,0 @@
-*.ts
-!*.d.ts
-
-# CDK asset staging directory
-.cdk.staging
-cdk.out
