--- lib/workload/stateless/stacks/icav2-copy-batch-utility/layers/pyproject.toml
@@ -1,32 +1,0 @@
-[build-system]
-requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
-
-[tool.poetry]
-name = "icav2_copy_batch_utility_tools"
-version = "0.0.1"
-description = "BSSH Manager Lambda Layers"
-license = "GPL-3.0-or-later"
-authors = [
-    "Alexis Lucattini"
-]
-homepage = "https://github.com/umccr/orcabus"
-repository = "https://github.com/umccr/orcabus"
-
-[tool.poetry.dependencies]
-python = "^3.11"
-boto3 = "^1.28"
-botocore = "^1.31"
-aws_requests_auth = "^0.4.3"
-wrapica = ">=1.0.0"
-
-[tool.poetry.group.dev]
-optional = true
-
-[tool.poetry.group.dev.dependencies]
-pyarrow = "^15.0.0"  # Pandas throws a warning if this is not installed
-pytest = "^7.0.0"  # For testing only
-# For typehinting only, not required at runtime
-mypy-boto3-ssm = "^1.34"
-mypy-boto3-secretsmanager = "^1.34"
-mypy-boto3-stepfunctions = "^1.34"
