--- lib/workload/stateless/stacks/icav2-copy-batch-utility/layers/src/icav2_copy_batch_utility_tools/utils/aws_ssm_helpers.py
@@ -1,63 +1,0 @@
-#!/usr/bin/env python
-
-"""
-Configuration handler
-
-Get ICAv2 configuration handler
-"""
-
-# Standard imports
-import typing
-import boto3
-from os import environ
-
-# Local imports
-from .globals import ICAV2_BASE_URL, ICAV2_ACCESS_TOKEN_URN_SSM_PATH
-
-# Type checking, only available as dev dependencies
-if typing.TYPE_CHECKING:
-    from mypy_boto3_ssm import SSMClient
-    from mypy_boto3_secretsmanager import SecretsManagerClient
-
-
-# AWS things
-def get_ssm_client() -> 'SSMClient':
-    """
-    Return SSM client
-    """
-    return boto3.client("ssm")
-
-
-def get_secrets_manager_client() -> 'SecretsManagerClient':
-    """
-    Return Secrets Manager client
-    """
-    return boto3.client("secretsmanager")
-
-
-def get_ssm_parameter_value(parameter_path) -> str:
-    """
-    Get the ssm parameter value from the parameter path
-    :param parameter_path:
-    :return:
-    """
-    return get_ssm_client().get_parameter(Name=parameter_path)["Parameter"]["Value"]
-
-
-def get_secret(secret_arn: str) -> str:
-    """
-    Return secret value
-    """
-    return get_secrets_manager_client().get_secret_value(SecretId=secret_arn)["SecretString"]
-
-
-# Set the icav2 environment variables
-def set_icav2_env_vars():
-    """
-    Set the icav2 environment variables
-    :return:
-    """
-    environ["ICAV2_BASE_URL"] = ICAV2_BASE_URL
-    environ["ICAV2_ACCESS_TOKEN"] = get_secret(
-        environ["ICAV2_ACCESS_TOKEN_SECRET_ID"]
-    )
